+++ android/src/main/kotlin/com/hodoan/custom_keyboard_action/KeyboardActionPlugin.kt
@@ -1,21 +1,11 @@
 package com.hodoan.custom_keyboard_action
 
 import android.annotation.SuppressLint
-import android.app.Activity
-import android.content.Context
-import android.content.Intent
-import android.graphics.PixelFormat
 import android.graphics.Rect
-import android.net.Uri
-import android.os.Build
 import android.os.Looper
-import android.provider.Settings
 import android.util.Log
-import android.view.*
-import android.widget.ImageButton
-import android.widget.TextView
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
+import android.view.View
+import android.view.ViewTreeObserver
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
@@ -32,10 +22,6 @@ class KeyboardActionPlugin : FlutterPlugin, MethodCallHandler,
     private lateinit var channel: MethodChannel
     private val keyEvent = KeyboardEvent()
     private var mainView: View? = null
-    private var viewToolBar: View? = null
-    private var windowManager: WindowManager? = null
-    private lateinit var context: Context
-    private lateinit var activity: Activity
 
     override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
         channel = MethodChannel(flutterPluginBinding.binaryMessenger, "custom_keyboard_action")
@@ -45,9 +31,6 @@ class KeyboardActionPlugin : FlutterPlugin, MethodCallHandler,
         ).setStreamHandler(
             keyEvent
         )
-        context = flutterPluginBinding.applicationContext
-        windowManager =
-            flutterPluginBinding.applicationContext.getSystemService(Context.WINDOW_SERVICE) as WindowManager?
         channel.setMethodCallHandler(this)
     }
 
@@ -65,72 +48,19 @@ class KeyboardActionPlugin : FlutterPlugin, MethodCallHandler,
 
     @SuppressLint("MissingInflatedId", "InflateParams")
     override fun onGlobalLayout() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            if (!Settings.canDrawOverlays(context)) {
-                val intent = Intent(
-                    Settings.ACTION_MANAGE_OVERLAY_PERMISSION,
-                    Uri.parse("package:" + activity.packageName)
-                )
-                activity.startActivityForResult(intent, 0)
-            }
-        }
-
         val r = Rect()
         mainView?.let {
             it.getWindowVisibleDisplayFrame(r)
 
-            val size = ViewCompat.getRootWindowInsets(it.rootView)
-            val heightBottom = size?.getInsets(WindowInsetsCompat.Type.ime())?.bottom ?: 0
-            val bottom = r.bottom - r.height()
-            val keyboardHeight = heightBottom - bottom
+//            val size = ViewCompat.getRootWindowInsets(it.rootView)
+//            val heightBottom = size?.getInsets(WindowInsetsCompat.Type.ime())?.bottom ?: 0
+//            val bottom = r.bottom - r.height()
+//            val keyboardHeight = heightBottom - bottom
 
             val screenHeight = it.rootView.height
             val keyPadHeight = screenHeight - r.bottom
             val newState = keyPadHeight > screenHeight * .15
-            if (newState) {
-                if (viewToolBar != null) {
-                    viewToolBar?.let { view ->
-                        windowManager?.removeView(view)
-                    }
-                    viewToolBar = null
-                }
-                @Suppress("DEPRECATION")
-                val layout = WindowManager.LayoutParams(
-                    ViewGroup.LayoutParams.MATCH_PARENT,
-                    ViewGroup.LayoutParams.WRAP_CONTENT,
-                    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O)
-                        WindowManager.LayoutParams.TYPE_SYSTEM_OVERLAY else
-                        WindowManager.LayoutParams.TYPE_APPLICATION_OVERLAY,
-                    WindowManager.LayoutParams.FLAG_NOT_FOCUSABLE,
-                    PixelFormat.TRANSLUCENT,
-                )
-                layout.gravity = Gravity.BOTTOM or Gravity.LEFT
-                val linearLayout =
-                    LayoutInflater.from(context).inflate(R.layout.custom_keyboard_action, null)
-                val doneBtn: TextView = linearLayout.findViewById(R.id.done)
-                doneBtn.setOnClickListener {
-                    keyEvent.success(ProtobufModel.ActionKeyboard.done)
-//                    val imm: InputMethodManager? =
-//                        activity.getSystemService(INPUT_METHOD_SERVICE) as InputMethodManager?
-//                    imm?.hideSoftInputFromWindow(activity.currentFocus?.windowToken, 0)
-                }
-                val btnDown: ImageButton = linearLayout.findViewById(R.id.down)
-                btnDown.setOnClickListener {
-                    keyEvent.success(ProtobufModel.ActionKeyboard.down)
-                }
-                val btnUp: ImageButton = linearLayout.findViewById(R.id.up)
-                btnUp.setOnClickListener {
-                    keyEvent.success(ProtobufModel.ActionKeyboard.up)
-                }
-                viewToolBar = linearLayout
-                windowManager?.addView(viewToolBar, layout)
-                viewToolBar?.setPadding(0, 0, 0, keyboardHeight)
-            } else {
-                viewToolBar?.let { view ->
-                    windowManager?.removeView(view)
-                }
-                viewToolBar = null
-            }
+            keyEvent.success(newState)
         }
     }
 
@@ -140,7 +70,6 @@ class KeyboardActionPlugin : FlutterPlugin, MethodCallHandler,
     }
 
     override fun onAttachedToActivity(binding: ActivityPluginBinding) {
-        activity = binding.activity
         try {
             mainView = binding.activity.findViewById(android.R.id.content)
             mainView!!.viewTreeObserver.addOnGlobalLayoutListener(this)
@@ -171,7 +100,7 @@ class KeyboardEvent : EventChannel.StreamHandler {
         sink = null
     }
 
-    fun success(value: ProtobufModel.ActionKeyboard) {
-        android.os.Handler(Looper.getMainLooper()).post { sink?.success(value.ordinal) }
+    fun success(value: Boolean) {
+        android.os.Handler(Looper.getMainLooper()).post { sink?.success(value) }
     }
 }