+++ example/lib/main.dart
@@ -1,6 +1,5 @@
-import 'package:flutter/material.dart';
-
 import 'package:custom_keyboard_action/custom_keyboard_action.dart';
+import 'package:flutter/material.dart';
 
 void main() {
   runApp(const MyApp());
@@ -19,24 +18,37 @@ class _MyAppState extends State<MyApp> {
     super.initState();
   }
 
+  final key = GlobalKey();
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: Scaffold(
+      home: ScaffoldKeyboard(
+        resizeToAvoidBottomInset: true,
         appBar: AppBar(
           title: const Text('Plugin example app'),
         ),
-        body: Column(
-          children: [
-            const Center(
-              child: TextField(
-                keyboardType: TextInputType.number,
-              ),
-            ),
-            TextFormFieldCustomKeyboard(
-              controller: TextEditingController(),
-            ),
-          ],
+        body: const SingleChildScrollView(
+          child: Column(
+            children: [
+              // const Stack(
+              //   children: [
+              //     TextField(
+              //       keyboardType: TextInputType.number,
+              //     ),
+              //     Positioned.fill(
+              //       child: TextFormFieldActions(),
+              //     ),
+              //   ],
+              // ),
+              SizedBox(height: 500),
+              TextField(),
+              // TextFormFieldCustomKeyboard(
+              //   key: key,
+              //   controller: TextEditingController(),
+              // ),
+            ],
+          ),
         ),
       ),
     );