+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+
+* remove maxLines Widget TextFormFieldCustomKeyboard
+* add widget ScaffoldKeyboard
 ## 0.0.2
 
 * update maxLines Widget TextFormFieldCustomKeyboard